Add files via upload
Code to make plots of the cost as a function of the population size. Makes 3 plots for time%2500, time%5000 and time%7500. Every million timesteps, between 86 million and 95 timesteps, represents a datapoint in the plot. Mutants are also plotted. Geom_text is used to label runs by the seed. Datapoints are seperated via colour between runs with evolving mutation and no evolving mutation.